Apply more test work and bugfixing

+++ src/libtrace/libtrace.h
@@ -166,8 +166,8 @@ typedef struct trc_gl_buffer_rev_t {
     trc_data_t* data;
     
     bool mapped;
-    size_t map_offset;
-    size_t map_length;
+    uint64_t map_offset;
+    uint64_t map_length;
     uint map_access; //Bitfield of GL_MAP_*_BIT
 } trc_gl_buffer_rev_t;
